custom=functions-metadata: move parseTree() to avoid fs import
Move parseTree() to ./src/parseTree.ts keeping generateCustomFunctionMetadata() in generate.ts.

This is to ensure ScriptLab can import parseTree() without having to import fs package,
because there's no longer a polyfill from react-scripts.
