+++ packages/custom-functions-metadata/src/main.ts
@@ -1,4 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+export * from "./parseTree";
 export * from "./generate";
